POCs/RealTime-Analytics/requirements.txt
pandas
scikit-learn
joblib
kafka-python
matplotlib